feat(settings): ai-consent section migrated + redirect, remove standalone page (#1608)
